refactor: extract student and instructor tables into separate components
- Created AdminStudentSearchTableComponent with its own logic and services
- Created AdminInstructorSearchTableComponent with its own logic and services
- Removed inline table HTML from admin-search-page.component.html
- Replaced with new component selectors using @Input()/@Output() bindings
- Encapsulated state management within each child component
- Added spec files following AdminAccountSearchTableComponent pattern

Fixes #13825
