+++ src/web/app/pages-admin/admin-search-page/admin-instructor-search-table/admin-instructor-search-table.component.spec.ts
@@ -1,0 +1,27 @@
+import { provideHttpClient } from '@angular/common/http';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { AdminInstructorSearchTableComponent } from './admin-instructor-search-table.component';
+
+describe('AdminInstructorSearchTableComponent', () => {
+  let component: AdminInstructorSearchTableComponent;
+  let fixture: ComponentFixture<AdminInstructorSearchTableComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      imports: [BrowserAnimationsModule],
+      providers: [provideHttpClient(), provideHttpClientTesting()],
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(AdminInstructorSearchTableComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
