+++ src/web/app/pages-admin/admin-search-page/admin-student-search-table/admin-student-search-table.component.spec.ts
@@ -1,0 +1,27 @@
+import { provideHttpClient } from '@angular/common/http';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { AdminStudentSearchTableComponent } from './admin-student-search-table.component';
+
+describe('AdminStudentSearchTableComponent', () => {
+  let component: AdminStudentSearchTableComponent;
+  let fixture: ComponentFixture<AdminStudentSearchTableComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      imports: [BrowserAnimationsModule],
+      providers: [provideHttpClient(), provideHttpClientTesting()],
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(AdminStudentSearchTableComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
